Move state from Square to Board
Remove the constructor from Square—it no longer needs its own state because the values of the squares will be passed in with props from Board.

Add constructor to Board, give initial value of the squares array.

Add onClick event to the renderSquare button. onClick redirects to Board's new function 'handleClick(i)'.

Pass the props from Board to Square. Replaces this.state with this.props for any data.